better names + doc
